Revert "renamed module to SparkIP"
This reverts commit fbce19a6ed7575d1d5173cf44cdb9e50f72d6200.
