+++ src/main/scala/com/databricks115/IPTraits.scala
@@ -1,4 +1,4 @@
-package com.SparkIP
+package com.databricks115
 
 trait IPv4Traits {
     protected def IPv4SubnetToCIDR(subnet: String): Int = {
